refactor(output_writer): move files around to make them more generically applicable
