+++ core/src/output_writer/get_connectivity_values_unstructured_mesh.h
@@ -30,4 +30,4 @@ public:
 
 } // namespace OutputWriter
 
-#include "output_writer/paraview/get_connectivity_values_unstructured_mesh.tpp"
+#include "output_writer/get_connectivity_values_unstructured_mesh.tpp"
