+++ core/src/output_writer/get_connectivity_values_unstructured_mesh.tpp
@@ -1,4 +1,4 @@
-#include "output_writer/paraview/get_connectivity_values_unstructured_mesh.h"
+#include "output_writer/get_connectivity_values_unstructured_mesh.h"
 
 namespace OutputWriter {
