+++ core/src/output_writer/loop_get_geometry_field_nodal_values.tpp
@@ -1,14 +1,9 @@
-#include "output_writer/paraview/loop_collect_field_variables_names.h"
-
-#include "output_writer/paraview/paraview_writer.h"
-#include "easylogging++.h"
-#include "utility/vector_operators.h"
 #include <cstdlib>
 #include "field_variable/field_variable.h"
 
 namespace OutputWriter {
 
-namespace ParaviewLoopOverTuple {
+namespace LoopOverTuple {
 
 /** Static recursive loop from 0 to number of entries in the tuple
  * Loop body
@@ -175,5 +170,5 @@ getGeometryFieldNodalValues(
 
   return false; // do not break iteration
 }
-} // namespace ParaviewLoopOverTuple
+} // namespace LoopOverTuple
 } // namespace OutputWriter
