+++ core/src/output_writer/loop_get_nodal_values.tpp
@@ -1,13 +1,8 @@
-#include "output_writer/paraview/loop_collect_field_variables_names.h"
-
-#include "output_writer/paraview/paraview_writer.h"
-
 #include <cstdlib>
 #include "field_variable/field_variable.h"
 
 namespace OutputWriter {
-
-namespace ParaviewLoopOverTuple {
+namespace LoopOverTuple {
 
 /** Static recursive loop from 0 to number of entries in the tuple
  * Loop body
@@ -182,5 +177,5 @@ getNodalValues(CurrentFieldVariableType currentFieldVariable,
 
   return false; // do not break iteration
 }
-} // namespace ParaviewLoopOverTuple
+} // namespace LoopOverTuple
 } // namespace OutputWriter
