+++ core/src/output_writer/paraview/paraview.tpp
@@ -9,8 +9,6 @@
 #include "base64.h"
 
 #include "output_writer/paraview/loop_output.h"
-#include "output_writer/paraview/loop_collect_mesh_properties.h"
-#include "output_writer/paraview/poly_data_properties_for_mesh.h"
 #include "control/diagnostic_tool/performance_measurement.h"
 
 namespace OutputWriter {
