+++ core/src/output_writer/paraview/paraview_write_combined_file.cpp
@@ -9,8 +9,6 @@
 #include "base64.h"
 
 #include "output_writer/paraview/loop_output.h"
-#include "output_writer/paraview/loop_collect_mesh_properties.h"
-#include "output_writer/paraview/poly_data_properties_for_mesh.h"
 
 namespace OutputWriter {
