+++ core/src/output_writer/paraview/paraview_write_combined_file_1D.tpp
@@ -9,11 +9,11 @@
 #include "easylogging++.h"
 #include "base64.h"
 
+#include "output_writer/poly_data_properties_for_mesh.h"
+#include "output_writer/loop_collect_mesh_properties.h"
+#include "output_writer/loop_get_nodal_values.h"
+#include "output_writer/loop_get_geometry_field_nodal_values.h"
 #include "output_writer/paraview/loop_output.h"
-#include "output_writer/paraview/loop_collect_mesh_properties.h"
-#include "output_writer/paraview/loop_get_nodal_values.h"
-#include "output_writer/paraview/loop_get_geometry_field_nodal_values.h"
-#include "output_writer/paraview/poly_data_properties_for_mesh.h"
 #include "control/diagnostic_tool/performance_measurement.h"
 
 namespace OutputWriter {
@@ -32,8 +32,7 @@ void Paraview::writePolyDataFile(
 
     // collect the size data that is needed to compute offsets for parallel file
     // output
-    ParaviewLoopOverTuple::loopCollectMeshProperties<
-        FieldVariablesForOutputWriterType>(
+    LoopOverTuple::loopCollectMeshProperties<FieldVariablesForOutputWriterType>(
         fieldVariables, meshPropertiesPolyDataFile_, meshNamesVector);
 
     Control::PerformanceMeasurement::stop("durationParaview1DInit");
@@ -282,7 +281,7 @@ void Paraview::writePolyDataFile(
 
   // collect all data for the field variables, organized by field variable names
   std::map<std::string, std::vector<double>> fieldVariableValues;
-  ParaviewLoopOverTuple::loopGetNodalValues<FieldVariablesForOutputWriterType>(
+  LoopOverTuple::loopGetNodalValues<FieldVariablesForOutputWriterType>(
       fieldVariables, vtkPiece1D_.meshNamesCombinedMeshes, fieldVariableValues);
 
   assert(!fieldVariableValues.empty());
@@ -346,7 +345,7 @@ void Paraview::writePolyDataFile(
 
   // collect all data for the geometry field variable
   std::vector<double> geometryFieldValues;
-  ParaviewLoopOverTuple::loopGetGeometryFieldNodalValues<
+  LoopOverTuple::loopGetGeometryFieldNodalValues<
       FieldVariablesForOutputWriterType>(
       fieldVariables, vtkPiece1D_.meshNamesCombinedMeshes, geometryFieldValues);
