+++ core/src/output_writer/paraview/paraview_write_combined_file_2D3D.tpp
@@ -9,11 +9,11 @@
 #include "easylogging++.h"
 #include "base64.h"
 
+#include "output_writer/poly_data_properties_for_mesh.h"
+#include "output_writer/loop_collect_mesh_properties.h"
+#include "output_writer/loop_get_nodal_values.h"
+#include "output_writer/loop_get_geometry_field_nodal_values.h"
 #include "output_writer/paraview/loop_output.h"
-#include "output_writer/paraview/loop_collect_mesh_properties.h"
-#include "output_writer/paraview/loop_get_nodal_values.h"
-#include "output_writer/paraview/loop_get_geometry_field_nodal_values.h"
-#include "output_writer/paraview/poly_data_properties_for_mesh.h"
 #include "control/diagnostic_tool/performance_measurement.h"
 
 namespace OutputWriter {
@@ -43,8 +43,7 @@ void Paraview::writeCombinedUnstructuredGridFile(
 
     // collect the size data that is needed to compute offsets for parallel file
     // output
-    ParaviewLoopOverTuple::loopCollectMeshProperties<
-        FieldVariablesForOutputWriterType>(
+    LoopOverTuple::loopCollectMeshProperties<FieldVariablesForOutputWriterType>(
         fieldVariables, meshPropertiesUnstructuredGridFile_, meshNamesVector);
 
     Control::PerformanceMeasurement::stop("durationParaview3DInit");
@@ -618,7 +617,7 @@ void Paraview::writeCombinedUnstructuredGridFile(
 
   // collect all data for the field variables, organized by field variable names
   std::map<std::string, std::vector<double>> fieldVariableValues;
-  ParaviewLoopOverTuple::loopGetNodalValues<FieldVariablesForOutputWriterType>(
+  LoopOverTuple::loopGetNodalValues<FieldVariablesForOutputWriterType>(
       fieldVariables, meshNamesSet, fieldVariableValues);
 
   if (!meshPropertiesInitialized) {
@@ -697,7 +696,7 @@ void Paraview::writeCombinedUnstructuredGridFile(
 
   // collect all data for the geometry field variable
   std::vector<double> geometryFieldValues;
-  ParaviewLoopOverTuple::loopGetGeometryFieldNodalValues<
+  LoopOverTuple::loopGetGeometryFieldNodalValues<
       FieldVariablesForOutputWriterType>(fieldVariables, meshNamesSet,
                                          geometryFieldValues);
