+++ core/src/output_writer/paraview/paraview_write_combined_values.tpp
@@ -10,10 +10,6 @@
 #include "base64.h"
 
 #include "output_writer/paraview/loop_output.h"
-#include "output_writer/paraview/loop_collect_mesh_properties.h"
-#include "output_writer/paraview/loop_get_nodal_values.h"
-#include "output_writer/paraview/loop_get_geometry_field_nodal_values.h"
-#include "output_writer/paraview/poly_data_properties_for_mesh.h"
 #include "control/diagnostic_tool/performance_measurement.h"
 
 namespace OutputWriter {
