+++ core/src/output_writer/poly_data_properties_for_mesh.cpp
@@ -1,4 +1,4 @@
-#include "output_writer/paraview/poly_data_properties_for_mesh.h"
+#include "output_writer/poly_data_properties_for_mesh.h"
 
 #include <Python.h> // has to be the first included header
 #include <iostream>
@@ -16,19 +16,19 @@ std::ostream &operator<<(std::ostream &stream, PolyDataPropertiesForMesh rhs) {
          << ", nCellsLocal: " << rhs.nCellsLocal
          << ", nPointsGlobal: " << rhs.nPointsGlobal
          << ", nCellsGlobal: " << rhs.nCellsGlobal << ", pointDataArrays:";
-  for (auto pointDataArray : rhs.pointDataArrays) {
+  for (const auto &pointDataArray : rhs.pointDataArrays) {
     stream << " [" << pointDataArray.name << "," << pointDataArray.nComponents
            << ": ";
-    for (int componentNo = 0;
-         componentNo < pointDataArray.componentNames.size(); componentNo++)
-      stream << "\"" << pointDataArray.componentNames[componentNo] << "\" ";
+    for (const auto &componentName : pointDataArray.componentNames) {
+      stream << "\"" << componentName << "\" ";
+    }
     stream << "]";
   }
   stream << ", " << rhs.unstructuredMeshConnectivityValues.size()
          << ", unstructuredMeshConnectivityValues: ";
 
-  for (int i = 0; i < rhs.unstructuredMeshConnectivityValues.size(); i++) {
-    stream << rhs.unstructuredMeshConnectivityValues[i] << " ";
+  for (int v : rhs.unstructuredMeshConnectivityValues) {
+    stream << v << " ";
   }
 
   stream << ")";
