+++ core/src/output_writer/poly_data_properties_for_mesh.h
